requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.10.2

# Configuration
django-environ==0.4.0



# Forms
django-braces==1.11.0
django-crispy-forms==1.6.0

# Models
django-model-utils==2.6

# Images
Pillow==3.4.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.27.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.7

# Redis support
django-redis==4.5.0
redis>=2.10.5





# Your custom requirements go here
django-auth-ldap==1.2.8
pyexcel-ods3==0.3.0
reportlab==3.3.0
pylabels==1.2.1
